backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic-settings==2.7.1
anthropic==0.43.0
openai==1.82.0
python-dotenv==1.0.1
redis==5.2.1
authlib==1.4.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
